+++ src/vault.rs
